Update to v0.4.0: Rebrand to PDF Tools and enhance package metadata
- Update manifest.json to version 0.2
- Rename package from "pdf-filler" to "pdf-tools" in package.json and package-lock.json
- Update dependencies to latest versions
- Remove deprecated binary files and installation scripts
- Revise README for installation instructions and package details

This update reflects the new branding and improves overall package management.
