--- pdf-filler-mcp-share/install.sh
@@ -1,38 +1,0 @@
-#!/bin/bash
-
-echo "🚀 Installing PDF Filler MCP Server..."
-npm install
-echo "✅ Dependencies installed!"
-echo ""
-
-# Get the absolute path automatically
-FULL_PATH="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)/server/index.js"
-
-echo "🎯 COPY THIS EXACT TEXT to your ~/.cursor/mcp.json:"
-echo ""
-echo "==============================================="
-echo '{'
-echo '  "mcpServers": {'
-echo '    "pdf-filler": {'
-echo '      "command": "node",'
-echo "      \"args\": [\"$FULL_PATH\"]"
-echo '    }'
-echo '  }'
-echo '}'
-echo "==============================================="
-echo ""
-echo "📂 Your MCP config file is located at:"
-echo "   ~/.cursor/mcp.json"
-echo ""
-echo "💡 To open it quickly:"
-echo "   Mac/Linux: open ~/.cursor/mcp.json"
-echo "   Or: code ~/.cursor/mcp.json"
-echo ""
-echo "📝 If you already have other MCP servers, just add the pdf-filler part inside your existing mcpServers section!"
-echo ""
-echo "🔄 After saving the file:"
-echo "1. Completely quit Cursor"
-echo "2. Restart Cursor" 
-echo "3. Look for 'pdf-filler' showing '10 tools enabled'"
-echo ""
-echo "✨ The full server path detected: $FULL_PATH"
