--- pdf-filler-mcp-share/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "pdf-filler",
-  "version": "0.3.0",
-  "description": "PDF filler MCP server for Claude Desktop",
-  "main": "server/index.js",
-  "license": "MIT",
-  "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.12.1",
-    "canvas": "^3.1.2",
-    "pdf-lib": "^1.17.1",
-    "pdf-parse": "^1.1.1",
-    "pdfjs-dist": "^2.16.105"
-  }
-}
